feat(core): 좌표계를 Vec3(float, 미터)로 전환하고 Sector/Zone 축을 XZ 로 재편
Week 1 의 1번 단계. 3D 전환의 밑바닥이라 다른 무엇보다 먼저 해야 한다 —
새 시스템을 2D short 로 먼저 짜면 전부 다시 쓰게 된다.

  Transform/Vec3.h   struct Vec3 { float x, y, z } + XZ 거리 (3장 규약)
  Transform          short m_x, m_y -> Vec3 m_pos + float m_yaw
  축 이름            ZoneCoord / SectorCoord / ZoneLayout / Sector 의
                     두 번째 축을 Y -> Z. 새 Y 는 높이이기 때문이다.
  월드 좌표 타입     Sector / SectorHelper / ZoneManager 를 short -> float

옛 GetY() 를 남기지 않고 삭제했다. 2D 시절의 y 는 높이가 아니라 지면의
두 번째 축이라, 이름만 두고 뜻을 바꾸면 놓친 호출부가 조용히 컴파일되면서
틀린다. 지우니 컴파일러가 전부 찾아냈다.

타입만 바꾸니 C4244(float -> short) 경고가 122건 떴다. 그대로 두면 어디가
의도한 절단이고 어디가 실수인지 구분할 수 없어 ToLegacyTile() 하나를 만들어
전부 그 문을 지나게 했다 (122건 암묵 -> 0건, 명시 호출 33곳). 아직 정수를
요구하는 곳은 패킷(5번 단계에서 교체)과 타일 격자(Week 2 Recast)뿐이다.

월드 크기와 패킷 레이아웃은 **일부러 그대로 두었다.** 그래야 기존
STRESS_TEST 가 회귀 검증 수단으로 남는다 — 실제로 소스 무수정 0 오류로
빌드됐고, 봇 1468마리가 Zone 16개에 고르게(86~103) 분포하며 몬스터를
1265마리 잡고 아이템 243개를 주웠다. 축이 바뀌었다면 분포가 쏠렸을 것이다.

이 커밋에는 작업 트리에 미커밋으로 남아 있던 Zone 이관 가드 작업
(Route.cpp 의 IsAllowedWhileTransferring, SectorHelper.cpp 주석)도 함께 들어간다.
같은 파일에서 좌표 변경과 줄이 얽혀 있어 따로 떼지 않았다.

+++ GameServer/Thread/DBThread.cpp
@@ -138,7 +138,11 @@ void DBThread::ProcessEvent(const shared_ptr<DB_EVENT_BASE>& event)
 			// Skip password verification — dev prototype with no auth requirement
 			DB_USER_INFO userInfo = connection->ExtractUserInfo(e->name);
 			vector<DB_ITEM_INFO> items = connection->ExtractInventory(userInfo._playerId);
-			GZoneManager->EnqueueByWorld(userInfo._x, userInfo._y, [session = e->session, userInfo, items]()
+			// DB 는 아직 2D 정수 좌표(_x, _y)를 들고 있다. 여기서 미터로 올린다 —
+			// _y 는 높이가 아니라 지면의 두 번째 축이므로 z 로 간다.
+			// DB 스키마를 Vec3 로 바꾸는 것은 Week 1 의 6번 단계에 묶어 처리한다.
+			GZoneManager->EnqueueByWorld(static_cast<float>(userInfo._x), static_cast<float>(userInfo._y),
+				[session = e->session, userInfo, items]()
 			{
 				UserHelper::HandleGetUserInfo(session, userInfo, items);
 			});
